+++ lib/widgets/app_bar/dashboard_app_bar.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:lotti/beamer/beamer_delegates.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/themes/theme.dart';
 
@@ -18,12 +20,20 @@ class DashboardAppBar extends StatelessWidget with PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      backgroundColor: colorConfig().headerBgColor,
+      backgroundColor: Colors.white,
+      elevation: 0,
+      scrolledUnderElevation: 10,
+      titleSpacing: 0,
       title: Text(
         dashboard.name,
-        style: appBarTextStyle(),
+        style: appBarTextStyleNew(),
       ),
-      centerTitle: true,
+      leading: IconButton(
+        padding: const EdgeInsets.symmetric(horizontal: 10),
+        onPressed: dashboardsBeamerDelegate.beamBack,
+        icon: SvgPicture.asset('assets/icons/back.svg'),
+      ),
+      centerTitle: false,
     );
   }
 }